fixed position mismatch of birds, made slingshort render like other entities

+++ core/src/main/java/io/github/angrybirbs/entities/General.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.physics.box2d.World;
 public class General extends Pig {
     private int power;
 
-    public General(World world, TiledMapTile tile, int x, int y) {
+    public General(World world, TiledMapTile tile, float x, float y) {
         super(world,tile,x,y);
         this.power = 10;
     }
