+++ core/src/main/java/io/github/angrybirbs/entities/King.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.physics.box2d.World;
 public class King extends Pig {
     private int power;
 
-    public King(World world, TiledMapTile tile, int x, int y) {
+    public King(World world, TiledMapTile tile, float x, float y) {
         super(world,tile,x,y);
         this.power = 10;
     }
