+++ core/src/main/java/io/github/angrybirbs/entities/Yellow.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.physics.box2d.World;
 public class Yellow extends Bird {
     private int power;
 
-    public Yellow(World world, TiledMapTile tile, int x, int y) {
+    public Yellow(World world, TiledMapTile tile, float x, float y) {
         super(world,tile,x,y);
         this.power = 10;
     }
